Major refactoring completed!
- Device and resources now follow Rust style traits to create.
- Added reference counting to resources.
- Refactored geometry cache to store geometries and hold vertex and index buffers.
- Refactored upload heap to be more robust. Heap blocks can now be directly written into.
- Refactored the command queue to include submission groups and requesting free command buffers.
- Clearing deleted resources no longer forces the device to wait idle.
- Demo application now does programmable vertex pulling instead of vertex input attributes.

+++ core/public/inl/core/cursor.h
@@ -32,7 +32,7 @@ struct Cursor
 	std::array<void*, CursorState::MAX> states;
 };
 
-using cursor_handle = lib::opaque_handle<Cursor, uint32, std::numeric_limits<uint32>::max(), struct CursorContext>;
+using cursor_handle = lib::opaque_handle<Cursor, uint64, std::numeric_limits<uint64>::max(), struct CursorContext>;
 
 struct CursorContext
 {
